Combine venues and qr-code routes, load config from params, trigger lambda via SQS
