+++ .env.sample
@@ -58,6 +58,10 @@ VERIFY_KEY_ID=
 VERIFY_PRIVATE_KEY=
 VERIFY_PUBLIC_KEY=
 
-EMAIL_ADDRESS_VERIFY_LIFETIME=60
+QR_EMAIL_ADDRESS_VERIFY_LIFETIME=60
+QR_APP_URL=
+QR_QUEUE_URL=
+QR_BUCKET_NAME=
+QR_SECRET=
 
 TIME_ZONE=UTC
