+++ lib/config/index.js
@@ -70,7 +70,11 @@ async function getConfig() {
       .prop('VERIFY_PRIVATE_KEY', S.string())
       .prop('VERIFY_PUBLIC_KEY', S.string())
       .prop('HSTS_MAX_AGE', S.number())
-      .prop('EMAIL_ADDRESS_VERIFY_LIFETIME', S.number())
+      .prop('QR_EMAIL_ADDRESS_VERIFY_LIFETIME', S.number())
+      .prop('QR_APP_URL', S.string())
+      .prop('QR_QUEUE_URL', S.string())
+      .prop('QR_BUCKET_NAME', S.string())
+      .prop('QR_SECRET', S.string())
   })
 
   const isProduction = /^\s*production\s*$/i.test(env.NODE_ENV)
@@ -172,8 +176,12 @@ async function getConfig() {
       privateKey: env.VERIFY_PRIVATE_KEY,
       publicKey: env.VERIFY_PUBLIC_KEY
     },
-    venues: {
-      emailAddressVerifyLifetime: env.EMAIL_ADDRESS_VERIFY_LIFETIME
+    qr: {
+      emailAddressVerifyLifetime: env.QR_EMAIL_ADDRESS_VERIFY_LIFETIME,
+      appUrl: env.QR_APP_URL,
+      queueUrl: env.QR_QUEUE_URL,
+      bucketName: env.QR_BUCKET_NAME,
+      secret: env.QR_SECRET
     }
   }
 
@@ -217,6 +225,7 @@ async function getConfig() {
     const jwtSecret = await getSecret('jwt')
     const deviceCheckSecret = await getSecret('device-check')
     const verifySecret = await getSecret('verify')
+    const qrSecret = await getSecret('qr')
 
     config.timeZone = await getParameter('time_zone')
     config.fastify.host = await getParameter('api_host')
@@ -269,9 +278,12 @@ async function getConfig() {
     config.aws.assetsBucket = await getParameter('s3_assets_bucket')
     config.aws.callbackQueueUrl = await getParameter('callback_url')
     config.metrics = JSON.parse(await getParameter('metrics_config'))
-    config.venues.emailAddressVerifyLifetime = Number(
+    config.qr.emailAddressVerifyLifetime = Number(
       await getParameter('email_address_verify_lifetime')
     )
+    config.qr.appUrl = await getParameter('qr_generate_url')
+    config.qr.queueUrl = await getParameter('qr_queue_url')
+    config.qr.bucketName = await getParameter('s3_qr_bucket')
 
     config.pgPlugin.read = await getParameter('db_reader_host')
     config.pgPlugin.write = await getParameter('db_host')
@@ -301,6 +313,7 @@ async function getConfig() {
     config.security.jwtSecret = jwtSecret.key
     config.deviceVerification = deviceCheckSecret
     config.verify = verifySecret
+    config.qr.secret = qrSecret.privateKey
   }
 
   return config
