--- lib/routes/venues/schema.js
@@ -1,67 +1,0 @@
-const S = require('fluent-schema')
-
-const email = {
-  body: S.object().prop(
-    'emailAddress',
-    S.string()
-      .format('email')
-      .required()
-  ),
-  response: {
-    204: S.null()
-  }
-}
-
-const risky = {
-  response: {
-    200: S.object().prop(
-      'venues',
-      S.array().items(
-        S.object()
-          .prop('id', S.string().required())
-          .prop(
-            'from',
-            S.string()
-              .format('date-time')
-              .required()
-          )
-          .prop(
-            'to',
-            S.string()
-              .format('date-time')
-              .required()
-          )
-      )
-    )
-  }
-}
-
-const verify = {
-  body: S.object()
-    .prop(
-      'emailAddress',
-      S.string()
-        .format('email')
-        .required()
-    )
-    .prop('verificationCode', S.string().required()),
-  response: {
-    200: S.object().prop('id', S.string().required())
-  }
-}
-
-const venueTypes = {
-  response: {
-    200: S.object().prop(
-      'venueTypes',
-      S.array().items(
-        S.object()
-          .prop('id', S.string().required())
-          .prop('name', S.string().required())
-          .prop('details', S.string().required())
-      )
-    )
-  }
-}
-
-module.exports = { email, risky, verify, venueTypes }
